+++ test/test-merge.js
@@ -18,7 +18,7 @@ const coverageOptions = {
     inline: true,
 
     // merge from exists raw dirs
-    inputDir: './docs/node-vm/raw, ./docs/cli/raw.zip, ./wrong-raw-dir',
+    inputDir: './docs/node-vm/raw, ./docs/cli/raw, ./wrong-raw-dir',
 
     name: 'My Merge Coverage Report',
     assetsPath: '../assets',